backend/requirements.txt
#Core libraries
langchain
llama-index

#For a backend server
fastapi
uvicorn

#Additional tools
requests
python-dotenv

#Optional AI agent frameworks
autogpt
babyagi